Tune hyperparameters for 83% linear probe accuracy
- Lower temperature from 0.5 to 0.2 for sharper hard-negative gradients
- Extend pretraining from 200 to 500 epochs
- Compute NT-Xent loss in float32 outside AMP autocast
- Enable cudnn.benchmark for faster convolution kernel selection
- Update README with results table and technical rationale

+++ src/pretrain.py
@@ -116,7 +116,7 @@ def pretrain(
             with torch.amp.autocast("cuda", enabled=(device.type == "cuda")):
                 _, z_i = model(x_i)
                 _, z_j = model(x_j)
-                loss = criterion(z_i, z_j)
+            loss = criterion(z_i.float(), z_j.float())
 
             scaler.scale(loss).backward()
             scaler.step(optimizer)
